requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cssselect==1.1.0
fake-useragent==0.1.11
feedparser==6.0.8
Flask==2.0.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.3
pandas==1.3.4
parse==1.19.0
prettytable==2.2.1
psycopg2-binary==2.9.1
pyecharts==1.9.0
pyee==8.2.2
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-html==0.10.0
scikit-learn==1.0
scipy==1.7.1
sgmllib3k==1.0.0
simplejson==3.17.5
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
threadpoolctl==3.0.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.2
yahoo-fin==0.8.9.1
zipp==3.6.0
